backend/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
asyncio==3.4.3
asyncpg==0.30.0
attrs==23.1.0
autobahn==19.11.2
Automat==22.10.0
autopep8==1.7.0
Babel==2.15.0
bcrypt==4.1.3
beautifulsoup4==4.12.3
billiard==4.2.0
bitarray==3.0.0
black==23.12.1
bleach==6.1.0
blinker==1.7.0
cachetools==5.2.1
certifi==2023.11.17
cffi==1.17.1
charset-normalizer==3.3.2
ckzg==2.0.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.2.2
comtypes==1.2.1
constantly==23.10.4
contourpy==1.0.5
cryptography==41.0.7
cycler==0.11.0
cytoolz==1.0.0
databases==0.9.0
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
dialogflow==2.0.0
ecdsa==0.19.0
eth-account==0.13.4
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.6.0
eth-rlp==2.1.0
eth-typing==5.0.1
eth-utils==5.1.0
eth_abi==5.1.0
executing==2.0.1
fastapi==0.115.2
fastjsonschema==2.19.1
flatbuffers==24.3.25
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.5.0
gast==0.4.0
gevent==24.10.3
google-api-core==1.34.0
google-api-python-client==2.72.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
google-cloud-speech==2.24.0
google-cloud-vision==3.6.0
google-pasta==0.2.0
googleapis-common-protos==1.58.0
greenlet==3.1.1
grpcio==1.60.1
grpcio-status==1.48.2
h11==0.14.0
h2==4.1.0
h5py==3.11.0
hexbytes==1.2.1
hpack==4.0.0
httpcore==1.0.5
httplib2==0.21.0
httpx==0.27.0
Hypercorn==0.17.3
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.10
incremental==22.10.0
iniconfig==2.0.0
ipykernel==6.29.4
ipython==8.24.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.25
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.8
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.1
keras==2.13.1
kiteconnect==5.0.0
kiwisolver==1.4.4
kombu==5.3.4
libclang==18.1.1
Markdown==3.6
MarkupSafe==2.1.3
matplotlib==3.6.1
matplotlib-inline==0.1.7
mibian==0.1.3
mistune==3.0.2
multidict==6.1.0
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
nltk==3.8.1
notebook==7.1.3
notebook_shim==0.2.4
numpy==2.1.2
oauthlib==3.2.2
onetimepass==1.0.1
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opt-einsum==3.3.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==2.2.3
pandocfilters==1.5.1
parsimonious==0.10.0
parso==0.8.4
passlib==1.7.4
pathspec==0.12.1
patsy==0.5.6
Pillow==9.2.0
platformdirs==4.1.0
pluggy==1.0.0
portalocker==2.8.2
priority==2.0.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
propcache==0.2.0
proto-plus==1.23.0
protobuf==3.20.3
psutil==5.9.8
psycopg2==2.9.10
pure-eval==0.2.2
py-solc-x==2.0.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAudio==0.2.14
pycodestyle==2.9.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.8.0
pymemcache==4.0.0
pyOpenSSL==23.3.0
pyparsing==3.0.9
pypiwin32==223
PySocks==1.7.1
pytest==7.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-json-logger==2.0.7
python-memcached==1.62
python-multipart==0.0.7
pyttsx3==2.90
pytz==2024.2
pyunormalize==16.0.0
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rlp==4.0.1
rpds-py==0.18.1
rsa==4.9
scipy==1.11.4
selenium==4.20.0
Send2Trash==1.8.3
service-identity==23.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.10.1
SQLAlchemy==2.0.36
stack-data==0.6.3
starlette==0.40.0
statsmodels==0.14.1
TA-Lib==0.4.28
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
terminado==0.18.1
tinycss2==1.3.0
toml==0.10.2
toolz==1.0.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.3
trio==0.25.0
trio-websocket==0.11.1
Twisted==23.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
types-python-dateutil==2.9.0.20240316
types-requests==2.32.0.20241016
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.1.0
uvicorn==0.25.0
vine==5.1.0
Wave==0.0.2
wcwidth==0.2.12
web3==7.4.0
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.0.1
wrapt==1.16.0
wsproto==1.2.0
yarl==1.17.1
zope.event==5.0
zope.interface==7.1.0
